third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_search-element_run_type=uri-expected.txt

This is a testharness.js-based test.
FAIL html5lib_search-element.html 114b9f3c8147c0ed8ef3ed1811a9da3f10d74402 assert_equals: expected "#document\n| <!DOCTYPE html>\n| <html>\n|   <head>\n|   <body>\n|     <p>\n|       \"foo\"\n|     <search>\n|       \"bar\"\n|       <p>\n|         \"baz\"" but got "#document\n| <!DOCTYPE html>\n| <html>\n|   <head>\n|   <body>\n|     <p>\n|       \"foo\"\n|       <search>\n|         \"bar\"\n|     <p>\n|       \"baz\""
FAIL html5lib_search-element.html 2204afb9037ec886f428ab5dcead5ee9f87c65cb assert_equals: expected "#document\n| <!DOCTYPE html>\n| <html>\n|   <head>\n|   <body>\n|     <search>\n|       <p>\n|         \"foo\"\n|     \"bar\"" but got "#document\n| <!DOCTYPE html>\n| <html>\n|   <head>\n|   <body>\n|     <search>\n|       <p>\n|         \"foobar\""
PASS html5lib_search-element.html 5153f797fbb63a23a40d19e298aca06d53d22f7f
Harness: the test ran to completion.
